added datalogger stuff

+++ FH17_EBIKE_SOFTWARE/ebike_startup.c
@@ -19,7 +19,7 @@
  * @param how a
  * @return a
  */
-int test(int what, double how){
+int test(int what, float how){
     return 0;
 }
